Fix lint and build warnings
