+++ src/components/IconButton.tsx
@@ -1,4 +1,4 @@
-import { ButtonHTMLAttributes } from 'react';
+import type { ButtonHTMLAttributes } from 'react';
 import './IconButton.css';
 
 export function IconButton({
