+++ src/components/ListItem.tsx
@@ -1,4 +1,4 @@
-import { HTMLAttributes } from 'react';
+import type { HTMLAttributes } from 'react';
 import './ListItem.css';
 
 export function ListItem({
